// All-Model-Chat-main/all-model-chat/types.ts
export interface ApiConfig {
  id: string;
  name: string;
  apiKey: string;
  apiProxyUrl?: string | null;
  isDefault?: boolean;
}